refactor LightConeEditor
