+++ libs/sr/ui/src/LightCone/LightConeEditor/index.tsx
@@ -1,1 +1,2 @@
 export * from './LightConeEditor'
+export * from './LightConeEditorCard'
